+++ daliuge-engine/dlg/drop_loaders.py
@@ -151,13 +151,12 @@ def save_binary(drop: "DataDROP", data: bytes):
     dropio.close()
 
 
-def load_utf8(drop: "DataDROP", allow_pickle=False):
+def load_utf8(drop: "DataDROP"):
     """
     Loads data from a drop and converts it to a UTF8 encoded string.
     """
     dropio = drop.getIO()
     dropio.open(OpenMode.OPEN_READ)
     res = dropio.buffer()
-    # res = str(io.BytesIO(dropio.buffer()), encoding="utf8")
     dropio.close()
     return res